226_Invert_Binary_Tree/main.go
package main

/*
https://leetcode.com/problems/invert-binary-tree/

Given the root of a binary tree, invert the tree, and return its root.

example 1
Input: root = [4,2,7,1,3,6,9]
Output: [4,7,2,9,6,3,1]

example 2
Input: root = [2,1,3]
Output: [2,3,1]

example 3
Input: root = []
Output: []
*/

func main() {

}

/**
 * Definition for a binary tree node.
 * type TreeNode struct {
 *     Val int
 *     Left *TreeNode
 *     Right *TreeNode
 * }
 */
type TreeNode struct {
	Val   int
	Left  *TreeNode
	Right *TreeNode
}

func invertTree(root *TreeNode) *TreeNode {
	if root == nil {
		return root
	}
	root.Left, root.Right = root.Right, root.Left
	invertTree(root.Left)
	invertTree(root.Right)
	return root
}
